requirements.txt
streamlit
numpy
faiss-cpu
requests